Updated RepeatBlocks and added CustomBlock and updated Tests
