+++ src/main/java/algoblocks/engine/CustomBlock.java
@@ -1,0 +1,22 @@
+package algoblocks.engine;
+
+public class CustomBlock extends EffectBlock{
+
+  public CustomBlock(){
+    super();
+  }
+
+  public Sequence getSequence() {
+    Sequence sequence = new Sequence();
+    for(Block block: blocks)
+      sequence.concatenate(block.getSequence());
+    return sequence;
+  }
+
+  public Sequence getInvertedSequence() {
+    Sequence sequence = new Sequence();
+    for(Block block: blocks)
+      sequence.concatenate(block.getInvertedSequence());
+    return sequence;
+  }
+}
